SMART AI PRO/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
bcrypt==4.0.1
spacy==3.7.2
nltk==3.8.1
wikipedia-api==0.6.0
sympy==1.12